Repetable common symlink script

+++ common/common_symlinks.sh
@@ -1,6 +1,27 @@
-ln -s $HOME/Developer/scaffold/common/dotfiles/gitconfig $HOME/.gitconfig
-ln -s $HOME/Developer/scaffold/common/dotfiles/gitignore $HOME/.gitignore
+#!/usr/bin/env bash
 
-ln -s $HOME/Developer/scaffold/common/fish $HOME/.config/fish
-ln -s $HOME/Developer/scaffold/common/emacs.d $HOME/.emacs.d
-ln -s $HOME/Developer/scaffold/common/weechat $HOME/.weechat
+if [ ! -L $HOME/.gitconfig ]; then
+  rm $HOME/.gitconfig 2> /dev/null
+  ln -s $HOME/Developer/scaffold/common/dotfiles/gitconfig $HOME/.gitconfig
+fi
+
+if [ ! -L $HOME/.gitignore ]; then
+  rm $HOME/.gitignore 2> /dev/null
+  ln -s $HOME/Developer/scaffold/common/dotfiles/gitignore $HOME/.gitignore
+fi
+
+if [ ! -L $HOME/.config/fish ]; then
+  rm -rf $HOME/.config/fish/ 2> /dev/null
+  mkdir -p $HOME/.config/
+  ln -s $HOME/Developer/scaffold/common/fish $HOME/.config/fish
+fi
+
+if [ ! -L $HOME/.emacs.d ]; then
+  rm -rf $HOME/.emacs.d/ 2> /dev/null
+  ln -s $HOME/Developer/scaffold/common/emacs.d $HOME/.emacs.d
+fi
+
+if [ ! -L $HOME/.weechat ]; then
+  rm -rf $HOME/.weechat/ 2> /dev/null
+  ln -s $HOME/Developer/scaffold/common/weechat $HOME/.weechat
+fi
